+++ src/eve-core/network/TCPConnection.cpp
@@ -327,7 +327,7 @@ bool TCPConnection::SendData( char* errbuf )
         mSendQueue.pop_front();
         mMSendQueue.Unlock();
 
-        int status = mSock->send( &(*buf)[ 0 ], (unsigned int)buf->size(), MSG_NOSIGNAL );
+        int status = mSock->send( &(*buf)[ 0 ], buf->size(), MSG_NOSIGNAL );
 
         if( status == SOCKET_ERROR )
         {
@@ -402,7 +402,7 @@ bool TCPConnection::RecvData( char* errbuf )
         else if( mRecvBuf->size() < TCPCONN_RECVBUF_SIZE )
             mRecvBuf->Resize<uint8>( TCPCONN_RECVBUF_SIZE );
 
-        int status = mSock->recv( &(*mRecvBuf)[ 0 ], (unsigned int)mRecvBuf->size(), 0 );
+        int status = mSock->recv( &(*mRecvBuf)[ 0 ], mRecvBuf->size(), 0 );
 
         if( status > 0 )
         {
